% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_iForest.R
\name{epi_iForest}
\alias{epi_iForest}
\title{Identifies epimutations using Isolation Forest}
\usage{
epi_iForest(mixture, case_id, ntrees)
}
\arguments{
\item{mixture}{beta values matrix. Samples in columns and
CpGs in rows.}

\item{case_id}{a character string specifying the name of the case sample.}

\item{ntrees}{number of binary trees to build for the model.
Default is \code{100}.}
}
\value{
The function returns the outlier score for the given case sample.
}
\description{
This function identifies regions with CpGs being outliers
using \link[isotree]{isolation.forest} approach.
}
