% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_isoforest.R
\name{epi_isoforest}
\alias{epi_isoforest}
\title{Identifies epimutations using Isolation Forest}
\usage{
epi_isoforest(mixture, case_id)
}
\arguments{
\item{mixture}{beta values matrix. Samples in columns and
CpGs in rows.}

\item{case_id}{a character string specifying the name of the case sample.}
}
\value{
The function returns the outlier score for the given case sample.
}
\description{
This function identifies regions with CpGs being outliers
using \link[isotree]{isolation.forest} approach.
}
