% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_mlm.R
\name{epi_mlm}
\alias{epi_mlm}
\title{Detects epimutations using Multivariate Linear Model (MLM)}
\usage{
epi_mlm(mixture, model)
}
\arguments{
\item{mixture}{beta values matrix.  Samples in columns and
CpGs in rows.}

\item{model}{design (or model) matrix.}
}
\value{
The function returns the F statistic,
R2 test statistic and Pillai.
}
\description{
Identifies CpGs with outlier methylation values
using methylated Multivariate Linear Model
}
