% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimutations.R
\name{epimutations}
\alias{epimutations}
\title{Epimutations analysis based on outlier detection methods}
\usage{
epimutations(
  case_samples,
  control_panel,
  method = "manova",
  chr = NULL,
  start = NULL,
  end = NULL,
  epi_params = epi_parameters(),
  bump_cutoff = 0.1,
  min_cpg = 3,
  verbose = TRUE
)
}
\arguments{
\item{case_samples}{a GenomicRatioSet object containing the case samples.
See the constructor function \link[minfi]{GenomicRatioSet}, \link[minfi]{makeGenomicRatioSetFromMatrix}.}

\item{control_panel}{a GenomicRatioSet object containing the control panel (control panel).
See the constructor function \link[minfi]{GenomicRatioSet}, \link[minfi]{makeGenomicRatioSetFromMatrix}.}

\item{method}{a character string naming the outlier detection method to be used.
This can be set as: \code{"manova"}, \code{"mlm"}, \code{"isoforest"}, \code{"mahdistmcd"},
\code{"barbosa"} and \code{"qn"}.
The default is \code{"manova"}.
For more information see \strong{Details}.}

\item{chr}{a character string containing the sequence names to be analysed. The default value is \code{NULL}.}

\item{start}{an integer specifying the start position. The default value is \code{NULL}.}

\item{end}{an integer specifying the end position. The default value is \code{NULL}.}

\item{epi_params}{The parameters for each method. See the function \link[epimutations]{epi_parameters}.}

\item{bump_cutoff}{a numeric value of the estimate of the genomic profile above the
cutoff or below the negative of the cutoff will be used as candidate regions.}

\item{min_cpg}{an integer specifying the minimum CpGs number in a DMR.}

\item{verbose}{logical. If TRUE additional details about the procedure will provide to the user.
The default is TRUE.}
}
\value{
The function returns an object of class tibble containing the outliers regions.
The results are composed by the following columns:
\itemize{
\item \code{epi_id}: systematic name for each epimutation identified. It provides the name
of the used anomaly detection method.
\item \code{sample}: the name of the sample containing the epimutation.
\item \code{chromosome}, \code{start} and \code{end}: indicate the location of the epimutation.
\item \code{sz}: the window's size of the event.
\item \code{cpg_n}: the number of CpGs in the epimutation.
\item \code{cpg_ids}: the names of CpGs in the epimutation.
\item \code{outlier_score}:
\itemize{
\item For method \code{manova} it provides the approximation to F-test and the Pillai score, separated by \code{/}.
\item For method \code{mlm} it provides the approximation to F-test and the R2 of the model, separated by \code{/}.
\item For method \code{isoforest} it provides the magnitude of the outlier score.
\item For methods \code{barbosa} and \code{mahdistmcd} it is filled with NA.
}
\item \code{outlier_significance}:
\itemize{
\item For methods \code{manova}, \code{mlm}, and \code{isoforest} it provides the p-value obtained from the model.
\item For method \code{barbosa} and \code{mahdistmcd} is filled with NA.
}
\item \code{outlier_direction}: indicates the direction of the outlier with \code{"hypomethylation"} and \code{"hypermethylation"}
\itemize{
\item For \code{manova}, \code{mlm}, \code{isoforest}, and \code{mahdistmcd} it is computed from the values obtained from bumphunter.
\item For \code{barbosa} it is computed from the location of the sample in the reference distribution (left vs. right outlier).
}
}
}
\description{
The function identifies  Differentially Methylated Regions
in a case sample by comparing it against a control panel.
}
\details{
The function compares a case sample against a control panel to identify epimutations in the given
sample. First, the DMRs are identified using the \link[bumphunter]{bumphunter} approach.
After that, CpGs in those DMRs are tested in order to detect regions
with CpGs being outliers.  For that,  different anomaly detection methods can be selected:
\itemize{
\item Multivariate Analysis of Variance (\code{"manova"}). \link[stats]{manova}
\item Multivariate Linear Model (\code{"mlm"})
\item Isolation Forest (\code{"isoforest"}) \link[isotree]{isolation.forest}
\item Robust Mahalanobis Distance (\code{"mahdistmcd"}) \link[robustbase]{covMcd}
\item Barbosa (\code{"barbosa"})
\item Qn (\code{"Qn"})
}
}
\examples{
\dontrun{
library(epimutacions)
data(methy)

#Find epimutations in GSM2562701 sample of methy dataset

case_sample <- methy[,"GSM2562701"]
control_panel <- methy[,-51]

epimutations(case_sample, control_panel, method = "manova")
}
}
