% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMutations.R
\name{format_bumps}
\alias{format_bumps}
\title{Convert bumps object to a tibble with additional formatting.}
\usage{
format_bumps(bumps, set, sample, reduced)
}
\arguments{
\item{bumps}{(bumps) Bump object returned by \link[bumphunter]{bumphunter}.}

\item{set}{(GenomicRatioSet, ExpressionSet) The dataset on which the bumps
were computed.}

\item{sample}{(string) The sample name for all the bumps.}

\item{reduced}{(bool) If True, only return a limited number of columns.
If False, return a full set of columns.}
}
\value{
A tibble dataframe where each row is an epimutation.
}
\description{
Convert bumps object to a tibble with additional formatting.
}
\keyword{internal}
