% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BumpBetaMatrix.R
\name{get_betas}
\alias{get_betas}
\title{Function to extract the methylation of a specific region}
\usage{
get_betas(bump, dataset)
}
\arguments{
\item{dataset}{The full methylation set as GenomicRatioSet or ExpressionSet.}

\item{bumps}{The region to be extracted as a bumps object.}
}
\value{
The betas as a matrix.
}
\description{
This function extracts the methylation from an GenomicRatioSet or
ExpressionSet object given a region defined as a bumps object.
}
