% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candRegsGR.R
\name{get_candRegsGR}
\alias{get_candRegsGR}
\title{Candidate regions to be epimutations}
\usage{
get_candRegsGR()
}
\value{
The function returns a GRanges object containing
the candidate regions.
}
\description{
Load candidate regions to be epimutations
from \code{epimutacionsData} package in \code{ExperimentHub}.
}
