% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ensemble_regulatory.R
\name{get_ENSEMBL_data}
\alias{get_ENSEMBL_data}
\title{Get ENSEMBL regulatory features overlapping a genomic region}
\usage{
get_ENSEMBL_data(chromosome, start, end, mart)
}
\arguments{
\item{chromosome}{Chromosome of the region}

\item{start}{Start of the region}

\item{end}{End of the region}

\item{mart}{\code{Mart} object to perform the ENSEMBL query}
}
\value{
\code{data.frame} of one row with the ENSEMBL regulatory
regions overlapping the genomic coordinate.
}
\description{
This function queries for ENSEMBL regulatory
features and collapse them to return
a single record.
}
