% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_beta.R
\name{getBetaParams}
\alias{getBetaParams}
\title{Model methylation as a beta distribution}
\usage{
getBetaParams(x)
}
\arguments{
\item{x}{Matrix of methylation expressed as a beta.
CpGs are in columns and samples in rows.}
}
\value{
Beta distribution.
}
\description{
Model methylation as a beta distribution
}
