% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRset.R
\docType{data}
\name{GRset}
\alias{GRset}
\title{GRset}
\format{
A GenomicRatioSet object with 4243 CpGs and 11 variables
}
\usage{
data(GRset)
}
\description{
A small GenomicRatioSet object to use in the functions examples
containing 10 control samples and a case sample.
}
\examples{
data(GRset)
}
\keyword{datasets}
