% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ensemble_regulatory.R
\name{merge_records}
\alias{merge_records}
\title{Merge records for the same ENSEMBL regulatory element}
\usage{
merge_records(tab)
}
\arguments{
\item{tab}{Results from \code{biomaRt::getBM} for the same regulatory element}
}
\value{
\code{data.frame} of one row after collapsing the
}
\description{
This function collapses the activity status of a
given an ENSEMBL regulatory
element in different tissues. Notice that tissues
identified as inactive will not be reported.
}
