% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methy.R
\docType{data}
\name{methy}
\alias{methy}
\title{\code{GenomicRatioSet} class object}
\format{
An object of class \code{GenomicRatioSet} with 80731 rows and 51 columns.
}
\usage{
data("methy")
}
\value{
A \code{GenomicRatioSet} object.
}
\description{
Dataset used to illustrate the package capabilities.
}
\examples{
data("methy")
sampleNames(methy)

}
\keyword{datasets}
