% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_matrix.R
\docType{data}
\name{methylation_matrix}
\alias{methylation_matrix}
\title{Matrix with simulated beta values for CpG islands
for testing purposes}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20 rows and 5 columns.
}
\usage{
data("methylation_matrix")
}
\value{
A matrix object.
}
\description{
The matrix contains 20 rows (CpG islands) and 5 columns
(5 samples). The beta values are uniformly distributed
between 0 and 0.1 for all samples and CpG islands.
Sample #1 values for CpG_4, _5, _6, _7, _8 are >> 0.1
}
\examples{
data("methylation_matrix")
dim(methylation_matrix)

}
\keyword{datasets}
