% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_pca_correction.R
\name{PCA_correction}
\alias{PCA_correction}
\title{PCA correction}
\usage{
PCA_correction(cases, controls, prange = 40000)
}
\arguments{
\item{cases}{a GenomicRatioSet object containing the case samples.
See the constructor function \link[minfi]{GenomicRatioSet},
\link[minfi]{makeGenomicRatioSetFromMatrix}.}

\item{controls}{a GenomicRatioSet object containing the
control panel (control panel).}

\item{prange}{probe.range}
}
\value{
A list with cases and controls corrected by PCs
}
\description{
Compute PCA correction
}
\examples{

data(res.epi.manova)
#Annotate the epimutations


}
