% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ensemble_regulatory.R
\name{process_ENSEMBL_results}
\alias{process_ENSEMBL_results}
\title{Process data from ENSEMBL}
\usage{
process_ENSEMBL_results(ensembl_res)
}
\arguments{
\item{ensembl_res}{Results from \code{biomaRt::getBM}}
}
\value{
\code{data.frame} of one row after collapsing
the input ENSEMBL regulatory regions
}
\description{
Process data from ENSEMBL to combine results
from the same regulatory elements in a unique record.
}
