% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.epi.manova.R
\docType{data}
\name{res.epi.manova}
\alias{res.epi.manova}
\title{res.epi.manova}
\format{
A data frame with 16 variables and 6 epimutations.
}
\usage{
data(res.epi.manova)
}
\value{
A data frame with 16 variables and 6 epimutations.
}
\description{
A data frame containing the results of \code{epimutations} function using
\code{"manova"} methods for \code{GRset} dataset. For more information
see the example of \code{epimutations} function.
}
\examples{
data(res.epi.manova)

}
\keyword{datasets}
