% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_isoforest.R
\name{res_isoforest}
\alias{res_isoforest}
\title{Creates a data frame containing the results
obtained from Isolation Forest}
\usage{
res_isoforest(bump, sts, outlier_score_cutoff)
}
\arguments{
\item{bump}{a DMR obtained from \link[bumphunter]{bumphunter}
(i.e. a row from \link[bumphunter]{bumphunter} method result).}

\item{sts}{the outlier score from
\link[epimutacions]{epi_isoforest} function results.}

\item{outlier_score_cutoff}{numeric specifying the outlier score cut off}

\item{beta_bump}{a beta values matrix for the CpGs in the selected
DMR. This matrix is the result of \link[epimutacions]{betas_from_bump}.}

\item{case}{a character string specifying the case sample name.}
}
\value{
The function returns a data frame containing the following information
for each DMR:
\itemize{
\item genomic ranges
\item DMR base pairs
\item number and name of CpGs in DMR
\item statistics:
\itemize{
\item Outlier score
\item Outlier significance
\item Outlier direction
}
\item Sample name
}

For more information about the output see \link[epimutacions]{epimutations}.
}
\description{
Creates a data frame containing the
genomic regions, statistics and direction for the DMRs.
}
