% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMutations.R
\name{run_bumphunter}
\alias{run_bumphunter}
\title{Run the Bumphunter algorithm.}
\usage{
run_bumphunter(set, design, ...)
}
\arguments{
\item{set}{A GenomicRatioSet or ExpressionSet}

\item{design}{Design matrix with rows representing samples and columns
representing covariates. Regression is applied to each row of set.}

\item{...}{Extra arguments to pass to \link[bumphunter]{bumphunter}}
}
\value{
The object returned by \link[bumphunter]{bumphunter}, or if no bumps
are found, a mock bumps object with zero rows returned by \code{empty_bumps}.
}
\description{
See \link[bumphunter]{bumphunter} doc for details.
}
\keyword{internal}
