% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cpg.R
\name{annotate_cpg}
\alias{annotate_cpg}
\title{Function to annotate DMR resulting from epimutacions package}
\usage{
annotate_cpg(
  data,
  db,
  split = ",",
  epi_col = "cpg_ids",
  gene_col = "GencodeBasicV12_NAME",
  feat_col = "Regulatory_Feature_Group",
  relat_col = "Relation_to_Island",
  build = "37",
  omim = T
)
}
\arguments{
\item{data}{DataFrame-like object.}

\item{db}{Database to use for annotation. E.g: IlluminaHumanMethylationEPICanno.ilm10b2.hg19.}

\item{split}{Separator for CpG ids. Default ','.}

\item{epi_col}{CpG ids, should be rownames in database.}

\item{gene_col}{Column name from where to extract gene names. Default: 'GencodeBasicV12_NAME'.}

\item{feat_col}{Column name from where to extract CpG feature groups. Default:  'Regulatory_Feature_Group'.}

\item{relat_col}{Column name from where to extract relation to island info. Default:  'Relation_to_Island'.}

\item{build}{Build for bioMart. Default '37'}

\item{omim}{Bool, if TRUE will annotate OMIMs as well. Takes a bit longer. Default TRUE}
}
\value{
DataFrame-like object annotated.
}
\description{
This function annotates a methylated region
}
