% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_epimutations.R
\name{annotate_epimutations}
\alias{annotate_epimutations}
\title{Annotate epimutations results}
\usage{
annotate_epimutations(
  epi_results,
  db = "IlluminaHumanMethylationEPICanno.ilm10b2.hg19",
  build = "37",
  ...
)
}
\arguments{
\item{epi_results}{Output from \code{epimutations}}

\item{db}{String with the Illumina annotation package used to annotate the CpGs}

\item{build}{Genomic build where the epimutations are mapped. By default, it is
GRCh37. Set build to \code{NULL} to use GRCh38.}

\item{...}{Further arguments passed to \code{annotate_CpG}}
}
\value{
Input object \code{epi_results} with additional columns with information about
the genes or overlapping regulatory features. See \code{annotate_cpg} and \code{add_ensemble_regulatory}
for an in-depth description of these variables.
}
\description{
Add information about close genes and regulatory elements for epimutations.
}
