% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_barbosa.R
\name{epi_barbosa}
\alias{epi_barbosa}
\title{Method implementing Barbosa et. al. 2019 to detect epimutations}
\usage{
epi_barbosa(
  case,
  fd,
  bctr_min,
  bctr_max,
  bctr_mean,
  bctr_pmin,
  bctr_pmax,
  window_sz = 1000,
  N = 3,
  offset_mean = 0.15,
  offset_abs = 0.1
)
}
\description{
The implementation of the method here is based on teh description of the
method used in:
Barbosa M, Joshi RS, Garg P, et al. Identification of rare de novo epigenetic
variations in congenital disorders. Nat Commun. 2018;9(1):2064. Published
2018 May 25. doi:10.1038/s41467-018-04540-x
}
