% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_mahdistmcd.R
\name{epi_mahdistmcd}
\alias{epi_mahdistmcd}
\title{Epimutation detection based on Outlier Detection with Robust Mahalanobis}
\usage{
epi_mahdistmcd(mixture, nsamp = c("best", "exact", "deterministic"))
}
\description{
The implementation of the method here is based on the discusion in this
thred of StakOverflow:
https://stats.stackexchange.com/questions/137263/multivariate-outlier-detection-with-robust-mahalanobis
}
