% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMANOVA.R
\name{epi_manova}
\alias{epi_manova}
\title{Identifies methylated outlier regions using manova}
\usage{
epi_manova(betas, model)
}
\arguments{
\item{betas}{Beta values matrix}

\item{Model}{Formula describing the model to be fitted}
}
\value{
F statistic
}
\description{
This function identifies regions with CpGs being outliers
}
