% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimutations.R
\name{epimutations}
\alias{epimutations}
\title{Epimutations analysis based on outlier detection methods}
\usage{
epimutations(
  case_samples,
  control_panel,
  method = "manova",
  chr = NULL,
  start = NULL,
  end = NULL,
  epi_params = epi_parameters(),
  bump_cutoff = 0.1,
  min_cpg = 3,
  verbose = TRUE
)
}
\arguments{
\item{case_samples}{a GenomicRatioSet object containing the case samples.
See the constructor function \link[minfi]{GenomicRatioSet}, \link[minfi]{makeGenomicRatioSetFromMatrix}.}

\item{control_panel}{a GenomicRatioSet object containing the control panel (control panel).
See the constructor function \link[minfi]{GenomicRatioSet}, \link[minfi]{makeGenomicRatioSetFromMatrix}.}

\item{method}{a character string naming the outlier detection method to be used.
This can be set as: \code{"manova"}, \code{"mlm"}, \code{"isoforest"}, \code{"mahdistmcd"},
\code{"barbosa"} and \code{"qn"}.
The default is \code{"manova"}.
For more information see \strong{Details}.}

\item{chr}{a character string containing the sequence names to be analysed. The default value is \code{NULL}.}

\item{start}{an integer specifying the start position. The default value is \code{NULL}.}

\item{end}{an integer specifying the end position. The default value is \code{NULL}.}

\item{epi_params}{The parameters for each method. See the function \link[epimutations]{epi_parameters}.}

\item{bump_cutoff}{a numeric value of the estimate of the genomic profile above the
cutoff or below the negative of the cutoff will be used as candidate regions.}

\item{min_cpg}{an integer specifying the minimum CpGs number in a DMR.}

\item{verbose}{logical. If TRUE additional details about the procedure will provide to the user.
The default is TRUE.}
}
\value{
The function returns an object of class tibble containing the outliers regions.
The results are composed by the following columns:
\itemize{
\item \code{epi_id}: the name of the anomaly detection method that has been used to detect the epimutation
\item \code{sample}: the name of the sample where the epimutation was found.
\item \code{chromosome}, \code{start} and \code{end}: the region coordinates.
\item \code{sz}: the number of base pairs in the region.
\item \code{cpg_n}: number of CpGs in the region.
\item \code{cpg_ids}: differentially methylated CpGs names.
\item \code{outlier_score}: the outlier score of the DMRs.
The outlier score only is available in the output of manova, mlm, isolation forest and qn.
\item \code{outlier_significance}: the outlier significance of the region.  Only it is
available for manova and mlm.
\item \code{outlier_direction}: describes if the epimutation is hypermethylated or hypomethylated.
The outlier direction is given by the methods barbosa, qn, isoforest, manova and mlm.
}
}
\description{
The function identifies  Differentially Methylated Regions
in a case sample by comparing it against a control panel.
}
\details{
The function compares a case sample against a control panel to identify epimutations in the given
sample. First, the DMRs are identified using the \link[bumphunter]{bumphunter} approach.
After that, CpGs in those DMRs are tested in order to detect regions
with CpGs being outliers.  For that,  different anomaly detection methods can be selected:
\itemize{
\item Multivariate Analysis of Variance (\code{"manova"}). \link[stats]{manova}
\item Multivariate Linear Model (\code{"mlm"})
\item Isolation Forest (\code{"isoforest"}) \link[isotree]{isolation.forest}
\item Robust Mahalanobis Distance (\code{"mahdistmcd"}) \link[robustbase]{covMcd}
\item Barbosa (\code{"barbosa"})
\item Qn (\code{"Qn"})
}
}
\examples{
library(epimutacions)
data(methy)

#Find epimutations in GSM2562701 sample of `methy` dataset

case_sample <- methy[,"GSM2562701"]
control_panel <- methy[,-51]

epimutations(case_sample, control_panel, method = "manova")
 
}
