% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiMutations.R
\name{EpiMutations}
\alias{EpiMutations}
\title{Return epimutations as per Aref-Eshghi et al, 2020.}
\usage{
EpiMutations(
  cases,
  controls,
  sample_id,
  cases_as_controls = T,
  args.bumphunter = list(cutoff = 0.1),
  num.cpgs = 10,
  pValue.cutoff = 0.01,
  outlier.score = 0.5,
  nsamp = "deterministic",
  method = c("manova", "mlm", "iso.forest", "Mahdist.MCD")
)
}
\arguments{
\item{cases}{(GenomicRatioSet, ExpressionSet) Case dataset.}

\item{controls}{(GenomicRatioSet, ExpressionSet) Control dataset.}

\item{sample_id}{(string) The column name in cases to compute epimutations for.}

\item{cases_as_controls}{(bool) If True, all remaining cases are added in controls.
If False, they are ignored.}

\item{args.bumphunter}{(list) Additional arguments to pass to
\link[bumphunter]{bumphunter}.}

\item{num.cpgs}{(integer) Bumps containing less cpgs than num.cpgs are discarded.}

\item{pValue.cutoff}{}

\item{outlier.score}{}

\item{nsamp}{}

\item{method}{(string) The outlier scoring method. Choose from
"manova", "mlm", "iso.forest", "Mahdist.MCD".}
}
\value{
A tibble of epimutation regions for sample_id.
}
\description{
Return epimutations as per Aref-Eshghi et al, 2020.
}
