% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicratioset.R
\docType{data}
\name{genomicratioset}
\alias{genomicratioset}
\title{\code{GenomicRatioSet} class object, created from
methylation_matrix dataset}
\format{
An object of class \code{GenomicRatioSet} with 20 rows and 5 columns.
}
\usage{
data("genomicratioset")
}
\value{
A \code{GenomicRatioSet} object.
}
\description{
\code{GenomicRatioSet} object, annotated with
\code{IlluminaHumanMethylation450kanno.ilmn12.hg19} data
with random sample choosing (via \code{sample(485512, 20)}).
Only rows for CpG_4, _5, _6, _7, _8 are chosen
with \code{rownames(Locations)[3000:3004]} in order to simulate
real hit with continous CpG island methylation difference.
}
\examples{
data("methylation_matrix")
dim(methylation_matrix)
sampleNames(genomicratioset)

}
\keyword{datasets}
