% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epi.R
\name{plot_epi}
\alias{plot_epi}
\title{Create DMRs and outlier CpGs plot}
\usage{
plot_epi(methy, epi_res, sam, chr, genome = "hg19", from = NULL, to = NULL)
}
\arguments{
\item{methy}{a \code{GenomicRatioSet} or \code{ExpressionSet}}

\item{sam}{A character specifying the sample to plot}

\item{chr}{A character specifying the chromosome of the epi-mutation to be plotted}

\item{genome}{The genome of reference.
It can be set as \code{"hg18"} and \code{"hg19"}. Default \code{"hg19"}.}

\item{from, to}{Scalar, specifying the range of genomic coordinates of the plot.
If \code{NULL} the plotting ranges are derived from the individual track.
Note that from cannot be larger than to.}
}
\description{
This function plots DMRs, outlier CpGs,
methylation in beta values and UCSC annotations for a specified epi-mutation
}
\details{
The tracks are plotted vertically. Each track is separate
}
